show specific user notes

+++ server/src/controllers/notes.ts
@@ -2,10 +2,13 @@ import { RequestHandler } from "express";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
 import NoteModel from "../models/note";
+import { assertIsDefined } from "../utils/assertIsDefined";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+  const authUserId = req.session.userId;
   try {
-    const notes = await NoteModel.find().exec();
+    assertIsDefined(authUserId)
+    const notes = await NoteModel.find({userId: authUserId}).exec();
     res.status(200).json(notes);
   } catch (error) {
     next(error);
@@ -14,13 +17,18 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getSingleNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authUserId = req.session.userId;
   try {
+    assertIsDefined(authUserId)
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note Id");
     }
     const note = await NoteModel.findById(noteId).exec();
     if (!note) {
       throw createHttpError(404, "Note not found");
+    }
+    if (!note.userId.equals(authUserId)){
+      throw createHttpError(401,"You cannot access this note!")
     }
     res.status(200).json(note);
   } catch (error) {
@@ -36,11 +44,15 @@ export const crateNote: RequestHandler<
 > = async (req, res, next) => {
   const title = req.body.title;
   const text = req.body.text;
+  const authUserId = req.session.userId;
   try {
+    assertIsDefined(authUserId)
+
     if (!title) {
       throw createHttpError(400, "Note must have a title!");
     }
     const newNote = await NoteModel.create({
+      userId: authUserId,
       title,
       text,
     });
@@ -59,7 +71,9 @@ export const updateNote: RequestHandler<
   const noteId = req.params.noteId;
   const newTitle = req.body.title;
   const newText = req.body.text;
+  const authUserId = req.session.userId;
   try {
+    assertIsDefined(authUserId)
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note Id");
     }
@@ -69,6 +83,9 @@ export const updateNote: RequestHandler<
     const note = await NoteModel.findById(noteId).exec();
     if (!note) {
       throw createHttpError(404, "Note not found");
+    }
+    if (!note.userId.equals(authUserId)){
+      throw createHttpError(401,"You cannot update this note!")
     }
     note.title = newTitle;
     note.text = newText;
@@ -83,13 +100,19 @@ export const updateNote: RequestHandler<
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authUserId = req.session.userId;
   try {
+    assertIsDefined(authUserId)
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note Id");
     }
     const note = await NoteModel.findById(noteId).exec();
     if (!note) {
       throw createHttpError(404, "Note not found");
+    }
+
+    if (!note.userId.equals(authUserId)){
+      throw createHttpError(401,"You cannot delete this note!")
     }
 
     await NoteModel.deleteOne({ _id: noteId });
